refactor: PasswordEncoder 인터페이스 분리 및 컨벤션 수정
