+++ src/main/java/shopping/application/AuthService.java
@@ -5,8 +5,8 @@ import shopping.auth.PasswordEncoder;
 import shopping.auth.TokenProvider;
 import shopping.domain.Email;
 import shopping.domain.User;
-import shopping.dto.LoginRequest;
-import shopping.dto.LoginResponse;
+import shopping.dto.request.LoginRequest;
+import shopping.dto.response.LoginResponse;
 import shopping.exception.PasswordNotMatchException;
 import shopping.exception.UserNotFoundException;
 import shopping.repository.UserRepository;
