+++ src/main/java/shopping/domain/Product.java
@@ -78,5 +78,4 @@ public class Product {
     public int hashCode() {
         return Objects.hash(id, name, imageUrl, price);
     }
-
 }
