+++ src/main/java/shopping/dto/request/CartItemUpdateRequest.java
@@ -1,4 +1,4 @@
-package shopping.dto;
+package shopping.dto.request;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
