+++ src/main/java/shopping/dto/request/ProductRequest.java
@@ -1,4 +1,4 @@
-package shopping.dto;
+package shopping.dto.request;
 
 import org.springframework.util.Assert;
