+++ src/test/java/shopping/application/CartServiceTest.java
@@ -18,8 +18,8 @@ import shopping.domain.CartItem;
 import shopping.domain.Email;
 import shopping.domain.Product;
 import shopping.domain.User;
-import shopping.dto.CartItemCreateRequest;
-import shopping.dto.CartItemUpdateRequest;
+import shopping.dto.request.CartItemCreateRequest;
+import shopping.dto.request.CartItemUpdateRequest;
 import shopping.exception.CartItemNotFoundException;
 import shopping.exception.UserNotFoundException;
 import shopping.exception.UserNotMatchException;
