+++ src/test/java/shopping/integration/AuthIntegrationTest.java
@@ -15,7 +15,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import shopping.auth.PasswordEncoder;
 import shopping.domain.User;
-import shopping.dto.LoginRequest;
+import shopping.dto.request.LoginRequest;
 import shopping.repository.UserRepository;
 
 class AuthIntegrationTest extends IntegrationTest {
